backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
groq==0.9.0
python-dotenv==1.0.1
pydantic==2.7.1
